Implement K Nearest neighbours

+++ AcademicProgressTracker/Models/Classification.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Value { get; set; }
-        public int LowerBound { get; set; }
-        public int UpperBound { get; set; }
+        public double LowerBound { get; set; }
+        public double UpperBound { get; set; }
     }
 }
